Data Structure and Algorithms 2.1
Write a Java class with a method that generates an array of 1 million
random integers. Add three methods to this class for bubble sort,
insertion sort and quick sort to use on this array. Time each algorithm
and compare the times. Draw a graph of time versus size of the array by
varying the elements of the array from starting from 10 in appropriate
regular steps for the algorithms and comment on the shape of the graph
in
relation to efficiency of each of the algorithms.